+++ src/ERC20UpgradeableTokenV1.sol
@@ -79,7 +79,6 @@ contract ERC20UpgradeableTokenV1 is
     function _authorizeUpgrade(address newImplementation) internal override onlyRole(UPGRADER_ROLE) {}
 
     // The following functions are overrides required by Solidity.
-
     function _update(address from, address to, uint256 value)
         internal
         override(ERC20Upgradeable, ERC20PausableUpgradeable)