+++ changelog.md
@@ -11,10 +11,16 @@
 
 ### Changed
 
+- Overall, we changed all the version of solidity into fixed 0.8.24.
+
 1. ERC20UpgradeableToken
 
 - Allow initializing the contract with a name, symbol, and burner. Including the originally generated code by Wizard of pauser, minter, and the default admin roles.
 - Modified the contract to add burner access right and burner to burn tokens.
+
+2. AmbassadorNft
+
+- Forked from OpenZeppelin's ERC1155 Wizard contract.
 
 ### Fixed